troubleshooting grafana repo

+++ release.sh
@@ -24,8 +24,15 @@ popd
 zip couchbase-datasource.zip "$tmp/couchbase-datasource" -r
 
 mkdir -p plugins/couchbase-datasource/versions/${CBVER}
-cp couchbase-datasource.zip plugins/couchbase-datasource/versions/${CBVER}
 
-cat plugins/repo/couchbase-datasource/index.html | jq ".versions+=[{\"arch\":{\"any\":{}},\"version\":\"${CBVER}\"}]" > "$tmp/cbds-versions.json"
-mv "$tmp/cbds-versions.json" plugins/repo/couchbase-datasource/index.html
-git add plugins
+if  [ ! -f plugins/couchbase-datasource/versions/${CBVER}/couchbase-datasource.zip ]; then
+  cp couchbase-datasource.zip plugins/couchbase-datasource/versions/${CBVER}
+
+  cat plugins/repo/index.html | jq ".plugins[0].versions+=[{\"version\":\"${CBVER}\"}]" > "$tmp/cb-repo-index.json"
+  mv "$tmp/cb-repo-index.json" plugins/repo/index.html
+
+  cat plugins/repo/couchbase-datasource/index.html | jq ".versions+=[{\"arch\":{\"any\":{}},\"version\":\"${CBVER}\"}]" > "$tmp/cbds-versions.json"
+  mv "$tmp/cbds-versions.json" plugins/repo/couchbase-datasource/index.html
+
+  git add plugins
+fi
